Fix #8
When "Send 1 ticket per quantity" is selected from the admin setting
then while sending the email it was over writting the last ticket
message with the whole email.

Where as it should attached all the created email message for each
quantity. I have not attached the whole email message content.