Fix #18
A single function creates the ticket content as well as generates the
ticket number and security code.

The ticket content generation was not working because of CRUD changes.
We now need to use get_<property> functions to get the data for the
items.

This has been modified for Woo 3.0.0. Backward compatibility has been
retained.

Ticket Number and Security code generation was not working as the order
object has also been modified in Woo 3.0.0. The same has been fixed in
this commit.

Tests should be run for all booking types in both versions of
WooCommerce.

The comments (from Tours Addon) printing in tickets should also be
tested.

Please note that WooCommerce Product Addons data which is printed in the
tickets will not work for WooCommerce 3.0.0. Tests need to be run and
the issue can be fixed accordingly. This can be run once Woo Product
Addons is updated for WooCommerce 3.0.0 compatibility.